+++ allink_core/core_apps/allink_teaser/cms_plugins.py
@@ -10,10 +10,19 @@ from allink_core.core.forms.fields import SelectLinkField
 from allink_core.core.forms.mixins import AllinkInternalLinkFieldMixin
 from allink_core.core_apps.allink_teaser.models import AllinkTeaserPlugin
 from allink_core.core.admin.mixins import AllinkMediaAdminMixin
+from allink_core.core.cms_plugins import CMSAllinkBaseSectionPlugin
+from .models import AllinkTeaserGridContainerPlugin
+
+
+@plugin_pool.register_plugin
+class CMSAllinkTeaserGridContainerPlugin(CMSAllinkBaseSectionPlugin):
+    model = AllinkTeaserGridContainerPlugin
+    name = 'Teaser Grid'
+    child_classes = ['CMSAllinkTeaserPlugin']
+    render_template = 'allink_teaser_grid/content.html'
 
 
 class AllinkTeaserPluginForm(AllinkInternalLinkFieldMixin, forms.ModelForm):
-
     internal_link = SelectLinkField(label='Link Internal', required=False)
 
     class Meta: