+++ CHANGELOG.md
@@ -83,6 +83,7 @@ general bugfixes
 - Optimized load more conditions and classes [#157](https://github.com/allink/allink-core/pull/157)
 - Used teaser alt text from original image [#158](https://github.com/allink/allink-core/pull/158)
 - Added list multicol cms plugin [#159](https://github.com/allink/allink-core/pull/159)
+- Added AllinkBaseSectionPlugin and AllinkTeaserGridContainerPlugin [#162](https://github.com/allink/allink-core/pull/162)
 - Added CONTENT_EXTENDED_FEATURE_SET feature toggle, which allows to hide functionality on the content plugin [#166](https://github.com/allink/allink-core/pull/166)
 - Updated info box plugin [#167](https://github.com/allink/allink-core/pull/167)
 - Added SVG Image plugin to generic plugin set [#169](https://github.com/allink/allink-core/pull/169)